+++ src/main/java/com/odos/odos_server/diary/Diary.java
@@ -1,6 +1,9 @@
-package com.odos.odos_server.entity;
+package com.odos.odos_server.diary;
 
 import com.odos.odos_server.Enum.Emotion;
+import com.odos.odos_server.challenge.Challenge;
+import com.odos.odos_server.challenge.Challenge_Diary;
+import com.odos.odos_server.member.Member;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
