+++ src/main/java/com/odos/odos_server/diary/Diary_Report.java
@@ -1,6 +1,7 @@
-package com.odos.odos_server.entity;
+package com.odos.odos_server.diary;
 
 import com.odos.odos_server.Enum.Report_Type;
+import com.odos.odos_server.member.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
