+++ src/main/java/com/odos/odos_server/interest/Interest.java
@@ -1,5 +1,7 @@
-package com.odos.odos_server.entity;
+package com.odos.odos_server.interest;
 
+import com.odos.odos_server.Enum.Interest_Type;
+import com.odos.odos_server.member.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,17 +11,17 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class Challenge_Like {
+public class Interest {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long diary_image_id;
+  private Long interest_id;
+
+  @Column
+  @Enumerated(EnumType.STRING)
+  private Interest_Type interest_category;
 
   @ManyToOne
   @JoinColumn(name = "member_id")
   private Member member;
-
-  @ManyToOne
-  @JoinColumn(name = "challenge_id")
-  private Challenge challenge;
 }
